Handle locked collections/items
Handle get/store/erase on locked collections and items. When an item is
locked, we typically get a "prompt" dbus object that we need to call the
"Prompt" method on. This causes the Secret Service (gnome-keyring) to
pop up a window prompting the user for the password to unlock the item.

Use a Go 1.23 range/iterator function to iterate over the search results
when looking up a secret. This handles the unlocked and locked results
transparently, unlocking any matching locked items found. The iterator
also returns only the items that match the attributes exactly. This
makes the implementation of Get and Delete simpler.

Bump the "go" version in `go.mod` so that we can use range functions.