Tidy up SecretService.Store
Tidy up the `SecretService.Store()` method to move the unlock logic to
the top of the function. This keeps code together - code for unlocking
and code for creating the item.

Reverse the test for prompting at the end of the function so it is the
same as the `Delete()` method - i.e. early return for the non-prompt
case. Keep code that does the same thing looking the same.