hermit: Upgrade gh
Upgrade gh from 2.54.0 to 2.55.0. This adds the `--editor`/`-e` flag to
`gh pr create` which is closer in user interface to `hub pull-request`.

This does not matter for this repository, but because this repo does use
`gh` for releases, I need to upgrade it for personal reasons as it gets
used whenever I run `gh` in this repo.

Gen-command: hermit upgrade gh