👷 ci: Split cicd workflow into ci and cd (#9)
Split the ci/cd workflow into separate CI and CD workflows, the former
of which only runs on PRs. The CD workflow calls the CI workflow as a
reusable workflow and does the release and howl-on-fail.

This is to simplify the pull requests on github so they only show the CI
job and not the other two as skipped.

Pull-request: https://github.com/foxygoat/git-credential-fdoss/pull/9