✨ Add encrypted sessions (#6)
Attempt to open an encrypted session when opening a session using the
`dh-ietf1024-sha256-aes128-cbc-pkcs7` algorithm specified in the Secret
Service specification. If we fail to open a session with that algorithm,
fall back to a `plain` session. Use the AES key negotiated when opening
the session to encrypt and decrypt the secrets exchanged with the Secret
Service.

The crypto routines in `crypto.go` have been taken from
https://github.com/mvdan/bitw - my thanks go to Daniel Martí for writing this
as it saved me a lot of headaches reading, understanding, implementing and
testing the algorithms as specified in the various RFCs.

A new dependency = `golang.org/x/crypto@v0.26.0` has been added, used by
the crypto routines needed to implemented encrypted sessions. Just the
`golang.org/x/crypto/hkdf` package of that module is used.

Turn off the `mnd` (magic number detector) linter as I prefer to decide
when magic numbers are appropriate or not, and not just cargo-cult the
idea that there should be none.

Issue: https://github.com/foxygoat/git-credential-fdoss/issues/4
Pull-request: https://github.com/foxygoat/git-credential-fdoss/pull/6