Fix links to specification
Fix the link to the specification in README.md which did not get fixed
when they recently changed the URL for the spec on freedesktop.org.

Update all the links in dbus.go to link to version 0.2 of the spec
instead of "latest" as perhaps they may one day publish a later version
and my links really should be for version 0.2.